+++ packages/eyes-sdk-core/lib/server/ServerConnector.js
@@ -165,7 +165,9 @@ class ServerConnector {
     const validStatusCodes = [HTTP_STATUS_CODES.OK, HTTP_STATUS_CODES.CREATED]
     if (validStatusCodes.includes(response.status)) {
       const runningSession = new RunningSession(response.data)
-      runningSession.setNewSession(response.status === HTTP_STATUS_CODES.CREATED)
+      if (response.data.isNew === undefined) {
+        runningSession.setIsNew(response.status === HTTP_STATUS_CODES.CREATED)
+      }
       this._logger.verbose('ServerConnector.startSession - post succeeded', runningSession)
       return runningSession
     }
@@ -192,7 +194,6 @@ class ServerConnector {
 
     const config = {
       name: 'stopSession',
-      isLongRequest: true,
       method: 'DELETE',
       url: GeneralUtils.urlConcat(
         this._configuration.getServerUrl(),
@@ -328,7 +329,6 @@ class ServerConnector {
 
     const config = {
       name: 'matchWindow',
-      isLongRequest: true,
       method: 'POST',
       url: GeneralUtils.urlConcat(
         this._configuration.getServerUrl(),
@@ -373,7 +373,6 @@ class ServerConnector {
 
     const config = {
       name: 'matchSingleWindow',
-      isLongRequest: true,
       method: 'POST',
       url: GeneralUtils.urlConcat(this._configuration.getServerUrl(), EYES_API_PATH),
       headers: {},
@@ -420,7 +419,6 @@ class ServerConnector {
 
     const config = {
       name: 'replaceWindow',
-      isLongRequest: true,
       method: 'PUT',
       url: GeneralUtils.urlConcat(
         this._configuration.getServerUrl(),