+++ packages/eyes-selenium/lib/EyesSelenium.js
@@ -799,13 +799,18 @@ class EyesSelenium extends Eyes {
    * @return {Promise<TestResults>}
    */
   async close(throwEx = true) {
+    let isErrorCaught = false
     const results = await super.close(true).catch(err => {
-      if (!throwEx) return err
-      else throw err
+      isErrorCaught = true
+      return err
     })
 
     if (this._runner) {
       this._runner._allTestResult.push(results)
+    }
+
+    if (throwEx && isErrorCaught) {
+      throw results
     }
 
     return results