+++ packages/eyes-webdriverio-4/index.js
@@ -27,3 +27,4 @@ exports.DeviceName = core.DeviceName
 exports.ScreenOrientation = core.ScreenOrientation
 exports.Region = core.Region
 exports.MatchLevel = core.MatchLevel
+exports.RectangleSize = core.RectangleSize