+++ packages/eyes-webdriverio-5/src/EyesVisualGrid.js
@@ -12,7 +12,6 @@ const {
   IgnoreRegionByRectangle,
   RectangleSize,
   Configuration,
-  TestResultsSummary,
   VisualGridRunner,
 } = require('@applitools/eyes-sdk-core')
 
@@ -42,6 +41,7 @@ class EyesVisualGrid extends EyesBase {
     super(serverUrl, isDisabled, new Configuration())
     this._runner = runner
     this._runner.attachEyes(this, this._serverConnector)
+    this._runner.makeGetVisualGridClient(makeVisualGridClient)
 
     /** @type {boolean} */ this._isOpen = false
     /** @type {boolean} */ this._isVisualGrid = true
@@ -51,7 +51,6 @@ class EyesVisualGrid extends EyesBase {
     /** @function */ this._checkWindowCommand = undefined
     /** @function */ this._closeCommand = undefined
     /** @function */ this._abortCommand = undefined
-    /** @type {Promise} */ this._closePromise = undefined
   }
 
   /**
@@ -109,7 +108,7 @@ class EyesVisualGrid extends EyesBase {
     if (this._runner.getConcurrentSessions())
       this._configuration.setConcurrentSessions(this._runner.getConcurrentSessions())
 
-    const {openEyes} = makeVisualGridClient({
+    const {openEyes} = await this._runner.getVisualGridClientWithCache({
       logger: this._logger,
       agentId: this.getFullAgentId(),
       apiKey: this._configuration.getApiKey(),
@@ -132,16 +131,8 @@ class EyesVisualGrid extends EyesBase {
     this._isOpen = true
 
     this._checkWindowCommand = checkWindow
-    this._closeCommand = async () => {
-      return close(true).catch(err => {
-        if (Array.isArray(err)) {
-          return err
-        }
-
-        throw err
-      })
-    }
-    this._abortCommand = async () => abort(true)
+    this._closeCommand = close
+    this._abortCommand = abort
 
     return this._driver
   }
@@ -160,37 +151,17 @@ class EyesVisualGrid extends EyesBase {
   }
 
   /**
-   * @param {boolean} [throwEx=true]
-   * @return {Promise<TestResults>}
+   * @return {Promise}
    */
-  async closeAndReturnResults(throwEx = true) {
-    try {
-      let resultsPromise = this._closePromise || this._closeCommand()
-      const res = await resultsPromise
-      const testResultSummary = new TestResultsSummary(res)
-
-      if (throwEx === true) {
-        for (const result of testResultSummary.getAllResults()) {
-          if (result.getException()) {
-            throw result.getException()
-          }
-        }
-      }
-
-      return testResultSummary
-    } finally {
-      this._isOpen = false
-      this._closePromise = undefined
-    }
+  async closeAsync() {
+    await this.close(false)
   }
 
   /**
    * @return {Promise}
    */
-  async closeAsync() {
-    if (!this._closePromise) {
-      this._closePromise = this._closeCommand()
-    }
+  async abortAsync() {
+    await this.abort()
   }
 
   /**
@@ -198,41 +169,31 @@ class EyesVisualGrid extends EyesBase {
    * @return {Promise<TestResults>}
    */
   async close(throwEx = true) {
-    const results = await this.closeAndReturnResults(throwEx)
+    let isErrorCaught = false
+    const results = await this._closeCommand(true).catch(err => {
+      isErrorCaught = true
+      return err
+    })
 
-    for (const result of results.getAllResults()) {
-      if (result.getException()) {
-        return result.getTestResults()
-      }
+    this._isOpen = false
+
+    if (this._runner) {
+      this._runner._allTestResult.push(...results)
     }
 
-    return results.getAllResults()[0].getTestResults()
-  }
+    if (throwEx && isErrorCaught) {
+      throw TypeUtils.isArray(results) ? results[0] : results
+    }
 
-  async abortIfNotClosed() {
-    return this.abort()
+    return results
   }
 
   /**
    * @return {Promise<?TestResults>}
    */
   async abort() {
-    if (typeof this._abortCommand === 'function') {
-      if (this._closePromise) {
-        this._logger.verbose('Can not abort while closing async, abort added to close promise.')
-        return this._closePromise.then(() => this._abortCommand(true))
-      }
-
-      return this._abortCommand()
-    }
-    return null
-  }
-
-  /**
-   * @return {Promise}
-   */
-  async abortAsync() {
-    this._closePromise = this.abort()
+    this._isOpen = false
+    return this._abortCommand()
   }
 
   /**
@@ -247,7 +208,7 @@ class EyesVisualGrid extends EyesBase {
    * @return {Promise<void>}
    */
   async closeAndPrintResults(throwEx = false) {
-    const results = await this.closeAndReturnResults(throwEx)
+    const results = await this.close(throwEx)
 
     const testResultsFormatter = new TestResultsFormatter(results)
     // eslint-disable-next-line no-console
@@ -293,9 +254,16 @@ class EyesVisualGrid extends EyesBase {
     }
 
     this._logger.verbose(`Dom extracted  (${checkSettings.toString()})   $$$$$$$$$$$$`)
-
     const source = await this._driver.getCurrentUrl()
-    const ignoreRegions = await this._prepareRegions(checkSettings.getIgnoreRegions())
+
+    const [ignore, floating, strict, layout, content, accessibility] = await Promise.all([
+      this._persistRegions(checkSettings.getIgnoreRegions()),
+      this._persistRegions(checkSettings.getFloatingRegions()),
+      this._persistRegions(checkSettings.getStrictRegions()),
+      this._persistRegions(checkSettings.getLayoutRegions()),
+      this._persistRegions(checkSettings.getContentRegions()),
+      this._persistRegions(checkSettings.getAccessibilityRegions()),
+    ])
 
     await this._checkWindowCommand({
       resourceUrls,
@@ -311,14 +279,23 @@ class EyesVisualGrid extends EyesBase {
       selector: targetSelector ? targetSelector : undefined,
       region: checkSettings.getTargetRegion(),
       scriptHooks: checkSettings.getScriptHooks(),
-      ignore: ignoreRegions,
-      floating: checkSettings.getFloatingRegions(),
+      ignore,
+      floating,
+      strict,
+      layout,
+      content,
+      accessibility,
       sendDom: checkSettings.getSendDom() ? checkSettings.getSendDom() : this.getSendDom(),
       matchLevel: checkSettings.getMatchLevel()
         ? checkSettings.getMatchLevel()
         : this.getMatchLevel(),
       source,
     })
+  }
+
+  async _persistRegions(regions) {
+    const persisted = await Promise.all(regions.map(r => r.toPersistedRegions(this._driver)))
+    return persisted.flat()
   }
 
   /**