+++ packages/eyes-webdriverio-5/src/fluent/IgnoreRegionByElement.js
@@ -1,6 +1,12 @@
 'use strict'
 
-const {GetRegion, Region, Location, CoordinatesType} = require('@applitools/eyes-sdk-core')
+const {
+  GetRegion,
+  Region,
+  Location,
+  CoordinatesType,
+  EyesJsBrowserUtils,
+} = require('@applitools/eyes-sdk-core')
 
 const {SelectorByElement} = require('./SelectorByElement')
 
@@ -38,6 +44,11 @@ class IgnoreRegionByElement extends GetRegion {
    */
   async getSelector(eyes) {
     return new SelectorByElement(this._element).getSelector(eyes)
+  }
+
+  async toPersistedRegions(driver) {
+    const xpath = await EyesJsBrowserUtils.getElementXpath(driver, this._element)
+    return [{type: 'xpath', selector: xpath}]
   }
 }