+++ packages/eyes-storybook/CHANGELOG.md
@@ -1,3 +1,21 @@
 # Changelog
 
 ## Unreleased
+
+
+## 3.2.14
+
+- Support both new and old server versions for identifying new running sessions. ([Trello](https://trello.com/c/mtSiheZ9/267-support-startsession-as-long-running-task))
+
+## 3.2.13
+
+- dom snapshot log for fetched resources now shows the byte length that was fetched [Trello](https://trello.com/c/CjSvn1OQ/262-storybook-409-conflict-wrong-sha)
+
+## 3.2.12
+
+- dom snapshot error while fetching is now always in clear text [Trello](https://trello.com/c/Jx1VJgpA/258-gap-storybook-assets-not-loading)
+
+## 3.2.11
+
+- fix trying to fetch branch info from server on non github integration runs
+- support future long running tasks [Trello](https://trello.com/c/60Rm4xXG/240-support-future-long-running-tasks)